add gradle build file
all dependencies are static and refer to jar files inside the libs/
directory in the project root. The latest version of the dependencies
can of course differ from the ones referenced. Maybe all dependencies
can be moved to maven central.

Be careful with the junixsocket dependency, as it requires some static
libraries.