add import adder class
this functionality of this class depends on the JAVASRC environment
variable to be set. It needs to point to the path of the javasrc zip
file (available with openjdk). This method is very sloppy, but I
recognized that there are not a lot of good ways to handle this task.